artistry-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
torch>=2.1.0
ultralytics>=8.2.0
opencv-python
pillow
numpy
timm
git+https://github.com/ChaoningZhang/MobileSAM.git
transformers>=4.41.0
accelerate
bitsandbytes
sentence-transformers
diffusers>=0.29.0
safetensors
xformers
pymongo
python-dotenv
